Zmiana komunikatu usuniecia usunietego boxa, uniemozliwienie wyswietlenia usunietego boxa, internacjonalizacja nazwy zwierzecia

+++ src/main/java/pl/lodz/p/it/ssbd2021/ssbd06/moh/endpoints/BoxEndpoint.java
@@ -165,6 +165,9 @@ public class BoxEndpoint extends AbstractEndpoint implements BoxEndpointLocal {
         if(isPending) {
             throw BoxException.boxIsPending();
         }
+        if (boxToDelete.isDelete()) {
+            throw BoxException.boxIsDeleted();
+        }
         boxManager.deleteBox(boxToDelete);
     }
